Refactor group home page tab views to separate components
