+++ app/app/home/group/types.ts
@@ -1,0 +1,6 @@
+import { NativeStackScreenProps } from "@react-navigation/native-stack";
+import { HomeStackParams } from "../types";
+
+export type GroupNavigationProps = {
+  navigation: NativeStackScreenProps<HomeStackParams, "group">["navigation"];
+};
